+++ R/dtwclust-classes.R
@@ -9,6 +9,7 @@
 #'   \item \code{distance}: A string indicating the distance used with \code{\link{dtwclust}}.
 #'   \item \code{centroid}: A string indicating the centroid used with \code{\link{dtwclust}}.
 #'   \item \code{preproc}: A string indicating the preprocessing used with \code{\link{dtwclust}}.
+#'   \item \code{datalist}: The provided data in the form of a list, where each element is a time series.
 #' }
 #'
 #' Additionally, the class inherits from \code{\link[flexclust]{kccasimple-class}}, so all related slots and
@@ -24,4 +25,5 @@ setClass("dtwclust", contains = c("kccasimple"),
          slots = c(type = "character",
                    distance = "character",
                    centroid = "character",
-                   preproc = "character"))
+                   preproc = "character",
+                   datalist = "list"))