+++ NEWS.md
@@ -1,5 +1,15 @@
 # NEWS
 
+## Version 1.0.0
+* Limited support for time series of different lengths
+* Improved plot method
+* More control parameters
+* Added more slots to the dtwclust class
+* Some parameters changed order
+* Default values of some auxiliary functions were changed (DBA and SBD)
+* No longer supporting native kccaFamilies
+* Several bug fixes, especially if 'dtw' or 'dtw2' were being used, in which case they may have been erroneously computed before
+
 ## Version 0.1.0
 
 * Initial release